+++ docs/CHANGELOG.rst
@@ -13,6 +13,9 @@ Unreleased
 Added
 -----
 * Add Mus musculus to all BCM workflows' schemas 
+* Add ``bam-split`` process with supporting processes 
+  ``upload-bam-primary``, ``upload-bam-secondary`` and
+  ``upload-header-sam``
 
 
 ==================